+++ CHANGELOG.md
@@ -2,6 +2,19 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+<a name="0.2.3"></a>
+
+### Bugfixes
+
+- Automatically fall back if scriptUrl is not specified
+- Automatically append trailing slash to trackerURL
+
+<a name="0.2.2"></a>
+
+### Features
+
+- Adds some error handling
+
 <a name="0.2.0"></a>
 
 ## [0.2.0](https://github.com/Arnaud73/ngx-matomo/compare/v0.1.4...v0.2.0) (2020-04-14)